layout: retire the NavBar cluster; auth pages go bare
The re-homed shell left most of containers/NavBar orphaned. The live pieces
move to their consumer: AuthState/useAuthState, RoleSwitcher, and the
Subscription license-state hook (its Display banner was dead — the rail's
Unlicensed item is the live affordance) relocate into containers/Sidebar
with their gql documents. Bare pages get a minimal BareHeader (midnight
band, home mark, visible focus ring) replacing NavBar.Container, and the
auth screens render bare instead of mounting the full shell. The NavBar
directory (NavBar, NavMenu, Controls, BucketSelect, Collaborators) is
deleted.
