--- catalog/app/containers/NavBar/BucketCollaborators.generated.ts
@@ -1,100 +1,0 @@
-/** Internal type. DO NOT USE DIRECTLY. */
-type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
-/** Internal type. DO NOT USE DIRECTLY. */
-export type Incremental<T> =
-  | T
-  | { [P in keyof T]?: P extends ' $fragmentName' | '__typename' ? T[P] : never }
-import type { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/core'
-import * as Types from '../../model/graphql/types.generated'
-
-export type BucketPermissionLevel = 'READ' | 'READ_WRITE'
-
-export type containers_NavBar_BucketCollaboratorsQueryVariables = Exact<{
-  bucket: string
-}>
-
-export interface containers_NavBar_BucketCollaboratorsQuery {
-  readonly __typename: 'Query'
-  readonly bucket: {
-    readonly __typename: 'Bucket'
-    readonly name: string
-    readonly collaborators: ReadonlyArray<{
-      readonly __typename: 'CollaboratorBucketConnection'
-      readonly permissionLevel: Types.BucketPermissionLevel
-      readonly collaborator: {
-        readonly __typename: 'Collaborator'
-        readonly email: string
-        readonly username: string
-      }
-    }>
-  } | null
-}
-
-export const containers_NavBar_BucketCollaboratorsDocument = {
-  kind: 'Document',
-  definitions: [
-    {
-      kind: 'OperationDefinition',
-      operation: 'query',
-      name: { kind: 'Name', value: 'containers_NavBar_BucketCollaborators' },
-      variableDefinitions: [
-        {
-          kind: 'VariableDefinition',
-          variable: { kind: 'Variable', name: { kind: 'Name', value: 'bucket' } },
-          type: {
-            kind: 'NonNullType',
-            type: { kind: 'NamedType', name: { kind: 'Name', value: 'String' } },
-          },
-        },
-      ],
-      selectionSet: {
-        kind: 'SelectionSet',
-        selections: [
-          {
-            kind: 'Field',
-            name: { kind: 'Name', value: 'bucket' },
-            arguments: [
-              {
-                kind: 'Argument',
-                name: { kind: 'Name', value: 'name' },
-                value: { kind: 'Variable', name: { kind: 'Name', value: 'bucket' } },
-              },
-            ],
-            selectionSet: {
-              kind: 'SelectionSet',
-              selections: [
-                { kind: 'Field', name: { kind: 'Name', value: 'name' } },
-                {
-                  kind: 'Field',
-                  name: { kind: 'Name', value: 'collaborators' },
-                  selectionSet: {
-                    kind: 'SelectionSet',
-                    selections: [
-                      {
-                        kind: 'Field',
-                        name: { kind: 'Name', value: 'collaborator' },
-                        selectionSet: {
-                          kind: 'SelectionSet',
-                          selections: [
-                            { kind: 'Field', name: { kind: 'Name', value: 'email' } },
-                            { kind: 'Field', name: { kind: 'Name', value: 'username' } },
-                          ],
-                        },
-                      },
-                      { kind: 'Field', name: { kind: 'Name', value: 'permissionLevel' } },
-                    ],
-                  },
-                },
-              ],
-            },
-          },
-        ],
-      },
-    },
-  ],
-} as unknown as DocumentNode<
-  containers_NavBar_BucketCollaboratorsQuery,
-  containers_NavBar_BucketCollaboratorsQueryVariables
->
-
-export { containers_NavBar_BucketCollaboratorsDocument as default }
