--- catalog/app/containers/NavBar/BucketCollaborators.graphql
@@ -1,12 +1,0 @@
-query ($bucket: String!) {
-  bucket(name: $bucket) {
-    name
-    collaborators {
-      collaborator {
-        email
-        username
-      }
-      permissionLevel
-    }
-  }
-}
