--- catalog/app/containers/NavBar/Collaborators.tsx
@@ -1,72 +1,0 @@
-import cx from 'classnames'
-import * as React from 'react'
-import * as M from '@material-ui/core'
-
-import { Avatars, Popup } from 'components/Collaborators'
-import * as style from 'constants/style'
-import type * as Model from 'model'
-import { useQueryS } from 'utils/GraphQL'
-import usePotentialCollaborators from 'utils/usePotentialCollaborators'
-
-import BUCKET_COLLABORATORS from './BucketCollaborators.generated'
-
-const NO_COLLABORATORS: ReadonlyArray<Model.GQLTypes.CollaboratorBucketConnection> = []
-
-const useStyles = M.makeStyles((t) => ({
-  avatars: {
-    marginLeft: t.spacing(2),
-    transition: 'opacity 0.3s ease',
-  },
-  hidden: {
-    opacity: 0,
-    zIndex: 0,
-  },
-}))
-
-interface CollaboratorsProps {
-  bucket: string
-  hidden: boolean
-}
-
-export default function Collaborators({ bucket, hidden }: CollaboratorsProps) {
-  const classes = useStyles()
-  const t = M.useTheme()
-  const sm = M.useMediaQuery(t.breakpoints.down('sm'))
-
-  const data = useQueryS(BUCKET_COLLABORATORS, { bucket })
-  const collaborators = data.bucket?.collaborators || NO_COLLABORATORS
-  const potentialCollaborators = usePotentialCollaborators()
-  const allCollaborators: Model.Collaborators = React.useMemo(
-    () => [
-      ...collaborators,
-      ...potentialCollaborators.map((collaborator) => ({
-        collaborator,
-        permissionLevel: undefined,
-      })),
-    ],
-    [collaborators, potentialCollaborators],
-  )
-
-  const [open, setOpen] = React.useState(false)
-  const handleOpen = React.useCallback(() => setOpen(true), [setOpen])
-  const handleClose = React.useCallback(() => setOpen(false), [setOpen])
-
-  if (!allCollaborators?.length) return null
-
-  return (
-    <M.MuiThemeProvider theme={style.appTheme}>
-      <Popup
-        bucket={bucket}
-        collaborators={allCollaborators}
-        onClose={handleClose}
-        open={open}
-      />
-      <Avatars
-        className={cx(classes.avatars, { [classes.hidden]: hidden })}
-        collaborators={allCollaborators}
-        iconized={sm}
-        onClick={handleOpen}
-      />
-    </M.MuiThemeProvider>
-  )
-}
