--- catalog/app/containers/NavBar/index.ts
@@ -1,1 +1,0 @@
-export * from './NavBar'
