+++ catalog/app/containers/Sidebar/RoleSwitcher.tsx
