+++ catalog/app/containers/Sidebar/gql/Subscription.graphql
