+++ catalog/app/containers/Sidebar/gql/SwitchRole.graphql
